Split node transfer filter helper
