+++ .pm/tasks/task_5ea9d9779665405face4c4984c056613.execution.md
@@ -235,6 +235,15 @@ Example:
 - Expected Result: the new proposal-time expiry regression passes, the new non-`main_token` asset regression passes, the full transfer-submit runtime suite stays green, the full `tests_action_payload` node module stays green, and diff whitespace remains clean.
 - Actual Result: passed. `tests_action_payload::pos_engine_proposal_filters_expired_transfer_actions_before_commit` passed (`1 passed`); `transfer_submit_api::tests::parse_transfer_submit_request_rejects_non_main_token_asset_id` passed (`1 passed`); runtime `transfer_submit` suite passed (`23 passed`); node `tests_action_payload` module passed (`40 passed`); `git diff --check` passed. All verification stayed on the canonical worktree shared cargo target path; package/build lock contention appeared repeatedly and was handled by waiting in place only, with no alternate compile directory.
 - Blocker / Next Action: inspect current branch diff, push the scoped fix onto `task/p2p-chain-transfer-submit-client`, then refresh PR #392 checks / review threads / mergeability and continue the watch-fix-merge loop.
+
+## 2026-06-09 21:24:12 CST / runtime_engineer
+- 完成内容: Repaired the fresh `required-gate` CI failure without changing transfer semantics by shrinking `crates/oasis7_node/src/node_engine_core.rs` back under the repository file-size gate.
+- 遗留事项: Push this structural follow-up commit, then wait for PR #392 checks to rerun and complete before merge.
+- Action: Moved the new proposal-time transfer filtering helpers out of `node_engine_core.rs` into a dedicated sibling module `crates/oasis7_node/src/node_engine_transfer_filter.rs`, kept `node_engine_core.rs` as the caller only, and wired the new module through `crates/oasis7_node/src/lib.rs`. This preserves the same `asset_id` / `valid_until_unix_ms` filtering behavior while bringing `node_engine_core.rs` from 1223 lines down to 1147 lines so `./scripts/check-rust-file-size.sh` passes.
+- Validation Command: `rtk wc -l crates/oasis7_node/src/node_engine_core.rs crates/oasis7_node/src/node_engine_transfer_filter.rs && rtk env -u RUSTC_WRAPPER cargo test -p oasis7_node tests_action_payload::pos_engine_proposal_filters_expired_transfer_actions_before_commit -- --nocapture && rtk ./scripts/check-rust-file-size.sh`
+- Expected Result: `node_engine_core.rs` falls below the 1200-line merge gate, the proposal-time expiry regression still passes after the helper extraction, and `check-rust-file-size.sh` reports no oversized code files.
+- Actual Result: passed. `node_engine_core.rs` measured `1147` lines and the extracted helper module `79`; `tests_action_payload::pos_engine_proposal_filters_expired_transfer_actions_before_commit` passed (`1 passed`); `check-rust-file-size.sh` reported `oversized code files=0` and `OK`.
+- Blocker / Next Action: commit and push the structural gate fix, then refresh PR #392 `required-gate` and continue the watch loop until all required checks are green and merge can proceed.
 - Review Evidence: runtime finding fixed and revalidated in the 2026-06-09 18:26:50 CST runtime entry above; QA wording/evidence findings disposition recorded in doc updates and prior review notes; producer/system-design scope decision on `fee_payer/sponsor/priority_fee` deferral recorded in the 2026-06-09 12:36:36 CST entry.
 - Review Findings Disposition: addressed
 - Finding Disposition Evidence: runtime reconciliation fix plus fresh shared-target verification chain; doc wording softened to keep residual doc-governance risk explicit rather than implied away.
