+++ crates/oasis7_node/src/lib.rs
@@ -55,6 +55,7 @@ mod node_engine_network;
 mod node_engine_replication;
 mod node_engine_slashing;
 mod node_engine_storage_challenge;
+mod node_engine_transfer_filter;
 mod node_runtime_core;
 mod pos_engine_gossip;
 mod pos_schedule;
